Fix AzureFile for uploading > 2^31 byte files

+++ karman-azure/src/main/groovy/com/bertramlabs/plugins/karman/azure/AzureFile.groovy
@@ -252,7 +252,7 @@ class AzureFile extends CloudFile {
 			ChunkedInputStream chunkedStream = new ChunkedInputStream(writeStream, partSize)
 			long filePosition = 0
 			int partNumber = 1
-			int startByte = 0
+			long startByte = 0
 			while(chunkedStream.available() >= 0 && (!contentLength || filePosition < contentLength)) {
 				// Last part can be less than 5 MB. Adjust part size.
